Break out a functional test case base class. Fix index controller since object registry change. Add functional test for index controller in order to catch it next time.
