+++ Tests/Controller/ResourceControllerTest.php
@@ -3,94 +3,23 @@
 namespace Lemon\RestBundle\Tests\Controller;
 
 use Doctrine\ORM\AbstractQuery;
+use Symfony\Component\HttpFoundation\Request;
+use Lemon\RestBundle\Tests\FunctionalTestCase;
 use Lemon\RestBundle\Event\RestEvents;
 use Lemon\RestBundle\Object\Criteria\DefaultCriteria;
 use Lemon\RestBundle\Object\Definition;
 use Lemon\RestBundle\Tests\Fixtures\Car;
 use Lemon\RestBundle\Tests\Fixtures\Tag;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-use Doctrine\ORM\Tools\SchemaTool;
 use Lemon\RestBundle\Tests\Fixtures\Person;
 use Lemon\RestBundle\Tests\Fixtures\FootballTeam;
 
-class ResourceControllerTest extends WebTestCase
+class ResourceControllerTest extends FunctionalTestCase
 {
-    /**
-     * @var \Symfony\Bundle\FrameworkBundle\Client
-     */
-    protected $client;
-    /**
-     * @var \Symfony\Component\DependencyInjection\ContainerInterface
-     */
-    protected $container;
-    /**
-     * @var \Doctrine\Bundle\DoctrineBundle\Registry
-     */
-    protected $doctrine;
-    /**
-     * @var \Doctrine\ORM\EntityManager
-     */
-    protected $em;
-    /**
-     * @var \JMS\Serializer\Serializer
-     */
-    protected $serializer;
-    /**
-     * @var \Lemon\RestBundle\Controller\ResourceController
-     */
-    protected $controller;
-
     public function setUp()
     {
-        $this->client = static::createClient();
-        $this->container = $this->client->getContainer();
-        $this->doctrine = $this->container->get('doctrine');
-        $this->em = $this->doctrine->getManager();
-        $this->serializer = $this->container->get('jms_serializer');
-
-        $schemaTool = new SchemaTool($this->em);
-        $schemaTool->createSchema($this->em->getMetadataFactory()->getAllMetadata());
-
-        $this->doctrine->getConnection()->beginTransaction();
-
-        $registry = $this->container->get('lemon_rest.object_registry');
-        $registry->add(new Definition('person', 'Lemon\RestBundle\Tests\Fixtures\Person'));
-        $registry->add(new Definition('footballTeam', 'Lemon\RestBundle\Tests\Fixtures\FootballTeam'));
+        parent::setUp();
 
         $this->controller = $this->container->get('lemon_rest.resource_controller');
-    }
-
-    protected static function getKernelClass()
-    {
-        return 'Lemon\RestBundle\Tests\TestKernel';
-    }
-
-    public function tearDown()
-    {
-        $this->doctrine->getConnection()->rollback();
-    }
-
-    /**
-     * @param string $method
-     * @param string $uri
-     * @param string|null $content
-     * @return Request
-     */
-    protected function makeRequest($method, $uri, $content = null, $parameters = array(), $server = array())
-    {
-        $request = Request::create(
-            $uri,
-            $method,
-            $parameters,
-            $cookies = array(),
-            $files = array(),
-            $server = array_merge(array(
-                'HTTP_ACCEPT' => 'application/json',
-            ), $server),
-            $content
-        );
-        return $request;
     }
 
     public function testListAction()
